(Fix): Fixed user_id on CustomerServices, added public/resources folder to store screen customization files and moved css and js to separate files
