+++ src/public/resources/css/customers/index.css
@@ -1,0 +1,26 @@
+.navbar-custom {
+    background-color: #ffffff;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    padding: 0.5rem 1rem;
+}
+
+.navbar-custom .navbar-brand {
+    color: #343a40;
+    height: 50px;
+    display: flex;
+    align-items: center;
+}
+
+.navbar-custom .navbar-nav .nav-item .nav-link {
+    padding: 0.5rem 1rem;
+}
+
+.navbar-custom .btn-outline-primary {
+    border-color: #007bff;
+    color: #007bff;
+}
+
+.navbar-custom .btn-outline-primary:hover {
+    background-color: #007bff;
+    color: #ffffff;
+}
